chore: update build process
Rename variables to match GNU standard practices when defining the C++
compiler, assembler, and linker to build with.

Refactor rules to look for source files in `./src` and header files in
`./lib`. Also separate rules for building the kernel binary and creating
the ISO image.

Utilise `bear` to generate a compilation database for `clangd`. This
will allow the Clang LSP to locate the headers in `./lib`.