chore(kernel): create global instance of `VgaWriter`
Declare variable `WRITER` to be used as the sole instance to write to
the VGA text buffer.

Set `VgaWriter` position with helper method `set_position` to easily
track places where `VgaWriter` position is mutated.

Remove use of `VgaWriter` in `call_ctors` as the writer instance gets
overidden by the call to its constructor in the for-loop. It may be
advisable to print to the serial port in this function instead.

Remove commented `global_ctors` function as it doesn't seem to be
invoked at all.