refactor(kernel): revert use of singleton pattern in `VgaWriter`
My previous implementation was "ill-formed", mostly due to the fact that
it was not thread-safe (even though there's no multithreading support
for this OS -- yet).

For now, I'm reverting to setting a global variable with an instance of
`VgaWriter`. Yes, this isn't any better, but it's good enough for a
kernel with currently no support for dynamic memory management. This
will call the default constructor, which interestingly enough doesn't
get invoked in `call_ctors`. Perhaps this initialization is inlined by
the linker? It doesn't seem to cause issues anyway.

I've also added a couple superfluous `KASSERT`s in the implementation of
`VgaWriter` because why not.