chore(drivers): mark all `VgaWriter` methods as `const`
Since printing to the VGA Text Mode buffer does not affect the way an
instance of this class shall be used, I've marked all member functions
as `const`. This clearly informs that the only side effect of using
these methods is that text will be displayed on screen. The fact that
the column and row positions are updated is irrelevant as these values
are not exposed in any way outside the class.

This has the added benefit that the `WRITER` instance can now be
declared as `const`, informing that it shall not be reassigned.