refactor(kernel): store 6 segment descriptors in the GDT
The GDT will now store 6 segment descriptors -- a null segment, two
kernel segments, two user segments, and one task-state segment (not yet
implemented). I've specified the same limit (4GiB) and similar flags for
all the kernel and user segments, as this is the recommended setup for
Flat/Long Mode GDTs, as documented in the OSDev Wiki.