chore: remove functionality to build ISO image
Currently, building the ISO image relies on GRUB Legacy. As the name
suggests, this software is no longer supported and is understandably
absent from many package managers today. This makes the process
cumbersome,  as booting the OS requires manually downloading unsupported
software.

Fortunately, the current build process produces a bootable ELF image,
which QEMU can boot directly. This eliminates the need for a bootable
ISO image, so I've removed this functionality altogether. In the future,
I may consider upgrading this process to GRUB 2 or Limine.